requirements.txt
# Core functionality
tavily-python==0.3.5
selenium==4.10.0
streamlit==1.25.0
openai==1.68.2
google-generativeai==0.8.4

# Development and utilities
python-dotenv==1.0.0      # For environment variables
webdriver-manager==4.0.1  # For Chrome driver management
loguru==0.7.2             # Better logging
pandas==2.2.0             # For data handling in Streamlit

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5    # For testing async functions 